\name{get_json_nearest}
\alias{get_json_nearest}
\title{Get JSON for schools}
\usage{
get_json_nearest(path, endpoints, query, n)
}
\arguments{
  \item{path}{URL path}

  \item{endpoints}{Node hierarchy}

  \item{query}{URL query}

  \item{n}{Number of rows to return}
}
\description{
Get JSON school data from the API.
}

