\name{get_json_schools}
\alias{get_json_schools}
\title{Get XML}
\usage{
get_json_schools(path, endpoints, query, n)
}
\arguments{
  \item{path}{URL path}

  \item{query}{URL query Get JSON

  Get JSON data from the API.}

  \item{path}{URL path}

  \item{query}{URL query Get JSON for schools

  Get JSON school data from the API.}

  \item{path}{URL path}

  \item{endpoints}{Node hierarchy}

  \item{query}{URL query}

  \item{n}{Number of rows to return}
}
\description{
Get XML data from the API.
}

