\name{GetNearestServiceUnit}
\alias{GetNearestServiceUnit}
\title{Get Service Unit data}
\usage{
GetNearestServiceUnit(unitType = 7, coords = c(6577574, 1627933), n = 5,
  apiKey = .serviceGuideKey, ...)
}
\description{
Get data from the Service Unit API
}
\examples{
\dontrun{
GetServiceUnit(endpoints = "ServiceUnitTypes")
}
}

