\name{GetServiceUnitSchools}
\alias{GetServiceUnitSchools}
\title{Get Service Unit data}
\usage{
GetServiceUnitSchools(endpoints = c("ServiceUnitTypeGroups", "7",
  "DetailedServiceUnits"), x = 6589642, y = 1620174,
  apiKey = .serviceGuideKey, ...)
}
\description{
Get data from the Service Unit API
}
\examples{
\dontrun{
GetServiceUnit(endpoints = "ServiceUnitTypes")
}
}

