\name{GetNearestServiceUnit}
\alias{GetNearestServiceUnit}
\title{Get data on the n nearest Service Units}
\usage{
GetNearestServiceUnit(unitType = 7, coords = c(6577574, 1627933), n = 5,
  apiKey = .unitKey, groups = TRUE, ...)
}
\description{
Get data on the n nearest Service Units from the Service
Unit API
}

